Add a fixup script to make enums compile correctly

+++ src/Yoga.NET.Interop/yoga/Align.cs
@@ -5,14 +5,14 @@ namespace Yoga.NET.Interop
     [NativeTypeName("uint8_t")]
     public enum Align : byte
     {
-        Auto = YGAlignAuto,
-        FlexStart = YGAlignFlexStart,
-        Center = YGAlignCenter,
-        FlexEnd = YGAlignFlexEnd,
-        Stretch = YGAlignStretch,
-        Baseline = YGAlignBaseline,
-        SpaceBetween = YGAlignSpaceBetween,
-        SpaceAround = YGAlignSpaceAround,
-        SpaceEvenly = YGAlignSpaceEvenly,
+        Auto = (byte)YGAlignAuto,
+        FlexStart = (byte)YGAlignFlexStart,
+        Center = (byte)YGAlignCenter,
+        FlexEnd = (byte)YGAlignFlexEnd,
+        Stretch = (byte)YGAlignStretch,
+        Baseline = (byte)YGAlignBaseline,
+        SpaceBetween = (byte)YGAlignSpaceBetween,
+        SpaceAround = (byte)YGAlignSpaceAround,
+        SpaceEvenly = (byte)YGAlignSpaceEvenly,
     }
 }
