+++ src/Yoga.NET.Interop/yoga/ExperimentalFeature.cs
@@ -5,6 +5,6 @@ namespace Yoga.NET.Interop
     [NativeTypeName("uint8_t")]
     public enum ExperimentalFeature : byte
     {
-        WebFlexBasis = YGExperimentalFeatureWebFlexBasis,
+        WebFlexBasis = (byte)YGExperimentalFeatureWebFlexBasis,
     }
 }
