+++ src/Yoga.NET.Interop/yoga/FlexDirection.cs
@@ -5,9 +5,9 @@ namespace Yoga.NET.Interop
     [NativeTypeName("uint8_t")]
     public enum FlexDirection : byte
     {
-        Column = YGFlexDirectionColumn,
-        ColumnReverse = YGFlexDirectionColumnReverse,
-        Row = YGFlexDirectionRow,
-        RowReverse = YGFlexDirectionRowReverse,
+        Column = (byte)YGFlexDirectionColumn,
+        ColumnReverse = (byte)YGFlexDirectionColumnReverse,
+        Row = (byte)YGFlexDirectionRow,
+        RowReverse = (byte)YGFlexDirectionRowReverse,
     }
 }
