+++ src/Yoga.NET.Interop/yoga/Unit.cs
@@ -5,9 +5,9 @@ namespace Yoga.NET.Interop
     [NativeTypeName("uint8_t")]
     public enum Unit : byte
     {
-        Undefined = YGUnitUndefined,
-        Point = YGUnitPoint,
-        Percent = YGUnitPercent,
-        Auto = YGUnitAuto,
+        Undefined = (byte)YGUnitUndefined,
+        Point = (byte)YGUnitPoint,
+        Percent = (byte)YGUnitPercent,
+        Auto = (byte)YGUnitAuto,
     }
 }
