+++ src/Yoga.NET.Interop/yoga/Wrap.cs
@@ -5,8 +5,8 @@ namespace Yoga.NET.Interop
     [NativeTypeName("uint8_t")]
     public enum Wrap : byte
     {
-        NoWrap = YGWrapNoWrap,
-        Wrap = YGWrapWrap,
-        WrapReverse = YGWrapWrapReverse,
+        NoWrap = (byte)YGWrapNoWrap,
+        Wrap = (byte)YGWrapWrap,
+        WrapReverse = (byte)YGWrapWrapReverse,
     }
 }
